refactor(chat): extract GroupProfileOverlay to own component
~1000 lines moved out of chat.tsx (now 4400 -> 3400 lines).

Extracted:
- src/components/GroupProfileOverlay.tsx (955 lines) — the overlay +
  its gp styles + formatRelative helper + BotDimensionChooser subcomp
- src/components/BotIcon.tsx (78 lines) — shared bot avatar (was used
  in 4 chat.tsx call-sites + overlay)
- src/hooks/useSwipeDismiss.ts (52 lines) — shared swipe-dismiss hook
  (chat.tsx still uses in 4 other overlays)

Wired up on explore tab:
- app/(tabs)/explore.tsx: selectedGroupId state + handleGroupPress +
  <GroupProfileOverlay /> render. Tap on feed item or avatar row now
  opens the overlay (member or non-member variant via isMember check).
- SearchGroupsModal closes first (200ms) then opens overlay.

Member/non-member behavior unchanged. RPC path, privacy gating, follow
hook integration, bot auto-save debounce all preserved.
